Use DeviceInterface callbacks in the Mlme
The MLME objects will use a DeviceInterface in their handler methods to
interact with external systems. This removes the dependency on the wlan
and ethernet interfaces in the Mlme and simplifies the Mlme object to
provide mostly routing.

The Device class is now primarily responsible for interfacing with the
"real world": event loops, timers, hardware, services, etc.

Change-Id: I3eea4a5582ecf64191728e26ec2b772b86f0d61a